refactor: move hints to features level
