+++ src/features/hints/index.ts
@@ -1,0 +1,8 @@
+export {
+  default as HintSettings,
+  type HintSettingsData,
+} from "./model/HintSettings";
+
+export { default as HintControls } from "./ui/HintControls";
+export { default as PronunciationHint } from "./ui/PronunciationHint";
+export { default as TranslationHint } from "./ui/TranslationHint";
