+++ src/features/hints/model/HintSettings.ts
