+++ src/features/hints/ui/TranslationHint.ts
@@ -1,6 +1,6 @@
 import { Component } from "../../../shared/ui";
 
-import RoundState from "../model/RoundState";
+import RoundState from "../../game/model/RoundState";
 import HintSettings from "../model/HintSettings";
 
 import { Observer, Publisher } from "../../../entities/state";
